partial rewrite using systemd and journald
This rewrite removes the use of subprocess, daemonize and other crates in favor of being a wrapper around init systems. This of course comes with a few drawbacks like: needing to add support for other init systems, a lot more checks and testing to make sure we know all kinds of responses we can get from different commands, making sure we don't leave random garbage around the scripts paths and probably more to be know.

With this change, there isn't a need for keeping cres processes open, running alongside the application anymore, which could lead to a lot of issues if there was an error in that process.

Alongside this rewrite, that I did not properly made commits for, hence why theres just one for all these changes, I changed the profile format from .json to .toml, .toml makes way more sense and is easier to understand. Theres no need for complex usage so I am not using the toml crate, this might change.

Though usable, this is very WIP.